+++ src/config/config_setup.h
@@ -3,7 +3,8 @@
     Gerbera - https://gerbera.io/
 
     config_setup.h - this file is part of Gerbera.
-    Copyright (C) 2020-2025 Gerbera Contributors
+
+    Copyright (C) 2020-2026 Gerbera Contributors
 
     Gerbera is free software; you can redistribute it and/or modify
     it under the terms of the GNU General Public License version 2