+++ src/config/grb_runtime.cc
@@ -3,7 +3,7 @@ Gerbera - https://gerbera.io/
 
     grb_runtime.cc - this file is part of Gerbera.
 
-    Copyright (C) 2024-2025 Gerbera Contributors
+    Copyright (C) 2024-2026 Gerbera Contributors
 
     Gerbera is free software; you can redistribute it and/or modify
     it under the terms of the GNU General Public License version 2
@@ -459,7 +459,7 @@ bool GerberaRuntime::addFile(const std::string& arg)
 bool GerberaRuntime::printCopyright(const std::string& arg)
 {
     fmt::print("\nGerbera UPnP Server {}\n"
-               "Copyright 2016-2025 Gerbera Contributors.\n"
+               "Copyright 2016-2026 Gerbera Contributors.\n"
                "Licence GPLv2: GNU GPL version 2.\n"
                "This is free software: you are free to change and redistribute it.\n\n",
         GERBERA_VERSION);
@@ -575,7 +575,7 @@ bool GerberaRuntime::printConfig()
 void GerberaRuntime::logCopyright()
 {
     log_info("Gerbera UPnP Server {} - {}", GERBERA_VERSION, DESC_MANUFACTURER_URL);
-    log_info("Copyright 2016-2025 Gerbera Contributors.");
+    log_info("Copyright 2016-2026 Gerbera Contributors.");
     log_info("Gerbera is free software, covered by the GNU General Public License version 2");
 }