+++ test/scripting/test_nfo_metafile.cc
@@ -4,7 +4,7 @@
 
     test_nfo_metafile.cc - this file is part of Gerbera.
 
-    Copyright (C) 2022-2025 Gerbera Contributors
+    Copyright (C) 2022-2026 Gerbera Contributors
 
     Gerbera is free software; you can redistribute it and/or modify
     it under the terms of the GNU General Public License version 2